+++ dotnet/src/Agents/OpenAI/OpenAIAssistantActions.cs
@@ -41,8 +41,6 @@ internal static class OpenAIAssistantActions
             RunStatus.Cancelled,
         ];
 
-    //private readonly Dictionary<string, ToolDefinition[]> _agentTools = [];
-
     /// <inheritdoc/> // %%%
     public static async Task CreateMessageAsync(AssistantsClient client, string threadId, ChatMessageContent message, CancellationToken cancellationToken)
     {
@@ -116,259 +114,259 @@ internal static class OpenAIAssistantActions
         while (messages.HasMore);
     }
 
-    //    /// <inheritdoc/> // %%%
-    //    public static async IAsyncEnumerable<ChatMessageContent> InvokeAsync(
-    //        OpenAIAssistantAgent agent,
-    //        [EnumeratorCancellation] CancellationToken cancellationToken)
-    //    {
-    //        if (agent.IsDeleted)
-    //        {
-    //            throw new KernelException($"Agent Failure - {nameof(OpenAIAssistantAgent)} agent is deleted: {agent.Id}.");
-    //        }
+    /// <inheritdoc/> // %%%
+    public static async IAsyncEnumerable<ChatMessageContent> InvokeAsync(
+        OpenAIAssistantAgent agent,
+        AssistantsClient client,
+        string threadId,
+        OpenAIAssistantConfiguration.PollingConfiguration pollingConfiguration,
+        ILogger logger,
+        [EnumeratorCancellation] CancellationToken cancellationToken)
+    {
+        if (agent.IsDeleted)
+        {
+            throw new KernelException($"Agent Failure - {nameof(OpenAIAssistantAgent)} agent is deleted: {agent.Id}.");
+        }
 
-    //        if (!this._agentTools.TryGetValue(agent.Id, out ToolDefinition[]? tools))
-    //        {
-    //            tools = [.. agent.Tools, .. agent.Kernel.Plugins.SelectMany(p => p.Select(f => f.ToToolDefinition(p.Name, FunctionDelimiter)))];
-    //            this._agentTools.Add(agent.Id, tools);
-    //        }
+        ToolDefinition[]? tools = [.. agent.Tools, .. agent.Kernel.Plugins.SelectMany(p => p.Select(f => f.ToToolDefinition(p.Name, FunctionDelimiter)))];
 
-    //        this.Logger.LogDebug("[{MethodName}] Creating run for agent/thrad: {AgentId}/{ThreadId}", nameof(InvokeAsync), agent.Id, threadId);
+        logger.LogDebug("[{MethodName}] Creating run for agent/thrad: {AgentId}/{ThreadId}", nameof(InvokeAsync), agent.Id, threadId);
 
-    //        CreateRunOptions options =
-    //            new(agent.Id)
-    //            {
-    //                OverrideInstructions = agent.Instructions,
-    //                OverrideTools = tools,
-    //            };
+        CreateRunOptions options =
+            new(agent.Id)
+            {
+                OverrideInstructions = agent.Instructions,
+                OverrideTools = tools,
+            };
 
-    //        // Create run
-    //        ThreadRun run = await client.CreateRunAsync(threadId, options, cancellationToken).ConfigureAwait(false);
+        // Create run
+        ThreadRun run = await client.CreateRunAsync(threadId, options, cancellationToken).ConfigureAwait(false);
 
-    //        this.Logger.LogInformation("[{MethodName}] Created run: {RunId}", nameof(InvokeAsync), run.Id);
+        logger.LogInformation("[{MethodName}] Created run: {RunId}", nameof(InvokeAsync), run.Id);
 
-    //        // Evaluate status and process steps and messages, as encountered.
-    //        HashSet<string> processedStepIds = [];
-    //        Dictionary<string, FunctionCallContent> functionSteps = [];
+        // Evaluate status and process steps and messages, as encountered.
+        HashSet<string> processedStepIds = [];
+        Dictionary<string, FunctionCallContent> functionSteps = [];
 
-    //        do
-    //        {
-    //            // Poll run and steps until actionable
-    //            PageableList<RunStep> steps = await PollRunStatusAsync().ConfigureAwait(false);
+        do
+        {
+            // Poll run and steps until actionable
+            PageableList<RunStep> steps = await PollRunStatusAsync().ConfigureAwait(false);
 
-    //            // Is in terminal state?
-    //            if (s_terminalStatuses.Contains(run.Status))
-    //            {
-    //                throw new KernelException($"Agent Failure - Run terminated: {run.Status} [{run.Id}]: {run.LastError?.Message ?? "Unknown"}");
-    //            }
+            // Is in terminal state?
+            if (s_terminalStatuses.Contains(run.Status))
+            {
+                throw new KernelException($"Agent Failure - Run terminated: {run.Status} [{run.Id}]: {run.LastError?.Message ?? "Unknown"}");
+            }
 
-    //            // Is tool action required?
-    //            if (run.Status == RunStatus.RequiresAction)
-    //            {
-    //                this.Logger.LogDebug("[{MethodName}] Processing run steps: {RunId}", nameof(InvokeAsync), run.Id);
+            // Is tool action required?
+            if (run.Status == RunStatus.RequiresAction)
+            {
+                logger.LogDebug("[{MethodName}] Processing run steps: {RunId}", nameof(InvokeAsync), run.Id);
 
-    //                // Execute functions in parallel and post results at once.
-    //                FunctionCallContent[] activeFunctionSteps = steps.Data.SelectMany(step => ParseFunctionStep(agent, step)).ToArray();
-    //                if (activeFunctionSteps.Length > 0)
-    //                {
-    //                    // Emit function-call content
-    //                    yield return GenerateFunctionCallContent(agent.GetName(), activeFunctionSteps);
+                // Execute functions in parallel and post results at once.
+                FunctionCallContent[] activeFunctionSteps = steps.Data.SelectMany(step => ParseFunctionStep(agent, step)).ToArray();
+                if (activeFunctionSteps.Length > 0)
+                {
+                    // Emit function-call content
+                    yield return GenerateFunctionCallContent(agent.GetName(), activeFunctionSteps);
 
-    //                    // Invoke functions for each tool-step
-    //                    IEnumerable<Task<FunctionResultContent>> functionResultTasks = ExecuteFunctionSteps(agent, activeFunctionSteps, cancellationToken);
+                    // Invoke functions for each tool-step
+                    IEnumerable<Task<FunctionResultContent>> functionResultTasks = ExecuteFunctionSteps(agent, activeFunctionSteps, cancellationToken);
 
-    //                    // Block for function results
-    //                    FunctionResultContent[] functionResults = await Task.WhenAll(functionResultTasks).ConfigureAwait(false);
+                    // Block for function results
+                    FunctionResultContent[] functionResults = await Task.WhenAll(functionResultTasks).ConfigureAwait(false);
 
-    //                    // Process tool output
-    //                    ToolOutput[] toolOutputs = GenerateToolOutputs(functionResults);
+                    // Process tool output
+                    ToolOutput[] toolOutputs = GenerateToolOutputs(functionResults);
 
-    //                    await client.SubmitToolOutputsToRunAsync(run, toolOutputs, cancellationToken).ConfigureAwait(false);
-    //                }
+                    await client.SubmitToolOutputsToRunAsync(run, toolOutputs, cancellationToken).ConfigureAwait(false);
+                }
 
-    //                if (this.Logger.IsEnabled(LogLevel.Information)) // Avoid boxing if not enabled
-    //                {
-    //                    this.Logger.LogInformation("[{MethodName}] Processed #{MessageCount} run steps: {RunId}", nameof(InvokeAsync), activeFunctionSteps.Length, run.Id);
-    //                }
-    //            }
+                if (logger.IsEnabled(LogLevel.Information)) // Avoid boxing if not enabled
+                {
+                    logger.LogInformation("[{MethodName}] Processed #{MessageCount} run steps: {RunId}", nameof(InvokeAsync), activeFunctionSteps.Length, run.Id);
+                }
+            }
 
-    //            // Enumerate completed messages
-    //            this.Logger.LogDebug("[{MethodName}] Processing run messages: {RunId}", nameof(InvokeAsync), run.Id);
+            // Enumerate completed messages
+            logger.LogDebug("[{MethodName}] Processing run messages: {RunId}", nameof(InvokeAsync), run.Id);
 
-    //            IEnumerable<RunStep> completedStepsToProcess =
-    //                steps
-    //                    .Where(s => s.CompletedAt.HasValue && !processedStepIds.Contains(s.Id))
-    //                    .OrderBy(s => s.CreatedAt);
+            IEnumerable<RunStep> completedStepsToProcess =
+                steps
+                    .Where(s => s.CompletedAt.HasValue && !processedStepIds.Contains(s.Id))
+                    .OrderBy(s => s.CreatedAt);
 
-    //            int messageCount = 0;
-    //            foreach (RunStep completedStep in completedStepsToProcess)
-    //            {
-    //                if (completedStep.Type.Equals(RunStepType.ToolCalls))
-    //                {
-    //                    RunStepToolCallDetails toolCallDetails = (RunStepToolCallDetails)completedStep.StepDetails;
+            int messageCount = 0;
+            foreach (RunStep completedStep in completedStepsToProcess)
+            {
+                if (completedStep.Type.Equals(RunStepType.ToolCalls))
+                {
+                    RunStepToolCallDetails toolCallDetails = (RunStepToolCallDetails)completedStep.StepDetails;
 
-    //                    foreach (RunStepToolCall toolCall in toolCallDetails.ToolCalls)
-    //                    {
-    //                        ChatMessageContent? content = null;
+                    foreach (RunStepToolCall toolCall in toolCallDetails.ToolCalls)
+                    {
+                        ChatMessageContent? content = null;
 
-    //                        // Process code-interpreter content
-    //                        if (toolCall is RunStepCodeInterpreterToolCall toolCodeInterpreter)
-    //                        {
-    //                            content = GenerateCodeInterpreterContent(agent.GetName(), toolCodeInterpreter);
-    //                        }
-    //                        // Process function result content
-    //                        else if (toolCall is RunStepFunctionToolCall toolFunction)
-    //                        {
-    //                            FunctionCallContent functionStep = functionSteps[toolFunction.Id]; // Function step always captured on invocation
-    //                            content = GenerateFunctionResultContent(agent.GetName(), functionStep, toolFunction.Output);
-    //                        }
+                        // Process code-interpreter content
+                        if (toolCall is RunStepCodeInterpreterToolCall toolCodeInterpreter)
+                        {
+                            content = GenerateCodeInterpreterContent(agent.GetName(), toolCodeInterpreter);
+                        }
+                        // Process function result content
+                        else if (toolCall is RunStepFunctionToolCall toolFunction)
+                        {
+                            FunctionCallContent functionStep = functionSteps[toolFunction.Id]; // Function step always captured on invocation
+                            content = GenerateFunctionResultContent(agent.GetName(), functionStep, toolFunction.Output);
+                        }
 
-    //                        if (content is not null)
-    //                        {
-    //                            ++messageCount;
+                        if (content is not null)
+                        {
+                            ++messageCount;
 
-    //                            yield return content;
-    //                        }
-    //                    }
-    //                }
-    //                else if (completedStep.Type.Equals(RunStepType.MessageCreation))
-    //                {
-    //                    RunStepMessageCreationDetails messageCreationDetails = (RunStepMessageCreationDetails)completedStep.StepDetails;
+                            yield return content;
+                        }
+                    }
+                }
+                else if (completedStep.Type.Equals(RunStepType.MessageCreation))
+                {
+                    RunStepMessageCreationDetails messageCreationDetails = (RunStepMessageCreationDetails)completedStep.StepDetails;
 
-    //                    // Retrieve the message
-    //                    ThreadMessage? message = await this.RetrieveMessageAsync(messageCreationDetails, cancellationToken).ConfigureAwait(false);
+                    // Retrieve the message
+                    ThreadMessage? message = await RetrieveMessageAsync(messageCreationDetails, cancellationToken).ConfigureAwait(false);
 
-    //                    if (message is not null)
-    //                    {
-    //                        AuthorRole role = new(message.Role.ToString());
+                    if (message is not null)
+                    {
+                        AuthorRole role = new(message.Role.ToString());
 
-    //                        foreach (MessageContent itemContent in message.ContentItems)
-    //                        {
-    //                            ChatMessageContent? content = null;
+                        foreach (MessageContent itemContent in message.ContentItems)
+                        {
+                            ChatMessageContent? content = null;
 
-    //                            // Process text content
-    //                            if (itemContent is MessageTextContent contentMessage)
-    //                            {
-    //                                content = GenerateTextMessageContent(agent.GetName(), role, contentMessage);
-    //                            }
-    //                            // Process image content
-    //                            else if (itemContent is MessageImageFileContent contentImage)
-    //                            {
-    //                                content = GenerateImageFileContent(agent.GetName(), role, contentImage);
-    //                            }
+                            // Process text content
+                            if (itemContent is MessageTextContent contentMessage)
+                            {
+                                content = GenerateTextMessageContent(agent.GetName(), role, contentMessage);
+                            }
+                            // Process image content
+                            else if (itemContent is MessageImageFileContent contentImage)
+                            {
+                                content = GenerateImageFileContent(agent.GetName(), role, contentImage);
+                            }
 
-    //                            if (content is not null)
-    //                            {
-    //                                ++messageCount;
+                            if (content is not null)
+                            {
+                                ++messageCount;
 
-    //                                yield return content;
-    //                            }
-    //                        }
-    //                    }
-    //                }
+                                yield return content;
+                            }
+                        }
+                    }
+                }
 
-    //                processedStepIds.Add(completedStep.Id);
-    //            }
+                processedStepIds.Add(completedStep.Id);
+            }
 
-    //            if (this.Logger.IsEnabled(LogLevel.Information)) // Avoid boxing if not enabled
-    //            {
-    //                this.Logger.LogInformation("[{MethodName}] Processed #{MessageCount} run messages: {RunId}", nameof(InvokeAsync), messageCount, run.Id);
-    //            }
-    //        }
-    //        while (RunStatus.Completed != run.Status);
+            if (logger.IsEnabled(LogLevel.Information)) // Avoid boxing if not enabled
+            {
+                logger.LogInformation("[{MethodName}] Processed #{MessageCount} run messages: {RunId}", nameof(InvokeAsync), messageCount, run.Id);
+            }
+        }
+        while (RunStatus.Completed != run.Status);
 
-    //        this.Logger.LogInformation("[{MethodName}] Completed run: {RunId}", nameof(InvokeAsync), run.Id);
+        logger.LogInformation("[{MethodName}] Completed run: {RunId}", nameof(InvokeAsync), run.Id);
 
-    //        // Local function to assist in run polling (participates in method closure).
-    //        async Task<PageableList<RunStep>> PollRunStatusAsync()
-    //        {
-    //            this.Logger.LogInformation("[{MethodName}] Polling run status: {RunId}", nameof(PollRunStatusAsync), run.Id);
+        // Local function to assist in run polling (participates in method closure).
+        async Task<PageableList<RunStep>> PollRunStatusAsync()
+        {
+            logger.LogInformation("[{MethodName}] Polling run status: {RunId}", nameof(PollRunStatusAsync), run.Id);
 
-    //            int count = 0;
+            int count = 0;
 
-    //            do
-    //            {
-    //                // Reduce polling frequency after a couple attempts
-    //                await Task.Delay(count >= 2 ? pollingConfiguration.RunPollingInterval : pollingConfiguration.RunPollingBackoff, cancellationToken).ConfigureAwait(false);
-    //                ++count;
+            do
+            {
+                // Reduce polling frequency after a couple attempts
+                await Task.Delay(count >= 2 ? pollingConfiguration.RunPollingInterval : pollingConfiguration.RunPollingBackoff, cancellationToken).ConfigureAwait(false);
+                ++count;
 
-    //#pragma warning disable CA1031 // Do not catch general exception types
-    //                try
-    //                {
-    //                    run = await client.GetRunAsync(threadId, run.Id, cancellationToken).ConfigureAwait(false);
-    //                }
-    //                catch
-    //                {
-    //                    // Retry anyway..
-    //                }
-    //#pragma warning restore CA1031 // Do not catch general exception types
-    //            }
-    //            while (s_pollingStatuses.Contains(run.Status));
+#pragma warning disable CA1031 // Do not catch general exception types
+                try
+                {
+                    run = await client.GetRunAsync(threadId, run.Id, cancellationToken).ConfigureAwait(false);
+                }
+                catch
+                {
+                    // Retry anyway..
+                }
+#pragma warning restore CA1031 // Do not catch general exception types
+            }
+            while (s_pollingStatuses.Contains(run.Status));
 
-    //            this.Logger.LogInformation("[{MethodName}] Run status is {RunStatus}: {RunId}", nameof(PollRunStatusAsync), run.Status, run.Id);
+            logger.LogInformation("[{MethodName}] Run status is {RunStatus}: {RunId}", nameof(PollRunStatusAsync), run.Status, run.Id);
 
-    //            return await client.GetRunStepsAsync(run, cancellationToken: cancellationToken).ConfigureAwait(false);
-    //        }
+            return await client.GetRunStepsAsync(run, cancellationToken: cancellationToken).ConfigureAwait(false);
+        }
 
-    //        // Local function to capture kernel function state for further processing (participates in method closure).
-    //        IEnumerable<FunctionCallContent> ParseFunctionStep(OpenAIAssistantAgent agent, RunStep step)
-    //        {
-    //            if (step.Status == RunStepStatus.InProgress && step.StepDetails is RunStepToolCallDetails callDetails)
-    //            {
-    //                foreach (RunStepFunctionToolCall toolCall in callDetails.ToolCalls.OfType<RunStepFunctionToolCall>())
-    //                {
-    //                    var nameParts = FunctionName.Parse(toolCall.Name, FunctionDelimiter);
+        // Local function to capture kernel function state for further processing (participates in method closure).
+        IEnumerable<FunctionCallContent> ParseFunctionStep(OpenAIAssistantAgent agent, RunStep step)
+        {
+            if (step.Status == RunStepStatus.InProgress && step.StepDetails is RunStepToolCallDetails callDetails)
+            {
+                foreach (RunStepFunctionToolCall toolCall in callDetails.ToolCalls.OfType<RunStepFunctionToolCall>())
+                {
+                    var nameParts = FunctionName.Parse(toolCall.Name, FunctionDelimiter);
 
-    //                    KernelArguments functionArguments = [];
-    //                    if (!string.IsNullOrWhiteSpace(toolCall.Arguments))
-    //                    {
-    //                        Dictionary<string, object> arguments = JsonSerializer.Deserialize<Dictionary<string, object>>(toolCall.Arguments)!;
-    //                        foreach (var argumentKvp in arguments)
-    //                        {
-    //                            functionArguments[argumentKvp.Key] = argumentKvp.Value.ToString();
-    //                        }
-    //                    }
+                    KernelArguments functionArguments = [];
+                    if (!string.IsNullOrWhiteSpace(toolCall.Arguments))
+                    {
+                        Dictionary<string, object> arguments = JsonSerializer.Deserialize<Dictionary<string, object>>(toolCall.Arguments)!;
+                        foreach (var argumentKvp in arguments)
+                        {
+                            functionArguments[argumentKvp.Key] = argumentKvp.Value.ToString();
+                        }
+                    }
 
-    //                    var content = new FunctionCallContent(nameParts.Name, nameParts.PluginName, toolCall.Id, functionArguments);
+                    var content = new FunctionCallContent(nameParts.Name, nameParts.PluginName, toolCall.Id, functionArguments);
 
-    //                    functionSteps.Add(toolCall.Id, content);
+                    functionSteps.Add(toolCall.Id, content);
 
-    //                    yield return content;
-    //                }
-    //            }
-    //        }
+                    yield return content;
+                }
+            }
+        }
 
-    //        async Task<ThreadMessage?> RetrieveMessageAsync(RunStepMessageCreationDetails detail, CancellationToken cancellationToken)
-    //        {
-    //            ThreadMessage? message = null;
+        async Task<ThreadMessage?> RetrieveMessageAsync(RunStepMessageCreationDetails detail, CancellationToken cancellationToken)
+        {
+            ThreadMessage? message = null;
 
-    //            bool retry = false;
-    //            int count = 0;
-    //            do
-    //            {
-    //                try
-    //                {
-    //                    message = await client.GetMessageAsync(threadId, detail.MessageCreation.MessageId, cancellationToken).ConfigureAwait(false);
-    //                }
-    //                catch (RequestFailedException exception)
-    //                {
-    //                    // Step has provided the message-id.  Retry on of NotFound/404 exists.
-    //                    // Extremely rarely there might be a synchronization issue between the
-    //                    // assistant response and message-service.
-    //                    retry = exception.Status == (int)HttpStatusCode.NotFound && count < 3;
-    //                }
+            bool retry = false;
+            int count = 0;
+            do
+            {
+                try
+                {
+                    message = await client.GetMessageAsync(threadId, detail.MessageCreation.MessageId, cancellationToken).ConfigureAwait(false);
+                }
+                catch (RequestFailedException exception)
+                {
+                    // Step has provided the message-id.  Retry on of NotFound/404 exists.
+                    // Extremely rarely there might be a synchronization issue between the
+                    // assistant response and message-service.
+                    retry = exception.Status == (int)HttpStatusCode.NotFound && count < 3;
+                }
 
-    //                if (retry)
-    //                {
-    //                    await Task.Delay(pollingConfiguration.MessageSynchronizationDelay, cancellationToken).ConfigureAwait(false);
-    //                }
+                if (retry)
+                {
+                    await Task.Delay(pollingConfiguration.MessageSynchronizationDelay, cancellationToken).ConfigureAwait(false);
+                }
 
-    //                ++count;
-    //            }
-    //            while (retry);
+                ++count;
+            }
+            while (retry);
 
-    //            return message;
-    //        }
-    //    }
+            return message;
+        }
+    }
 
     private static AnnotationContent GenerateAnnotationContent(MessageTextAnnotation annotation)
     {
