+++ dotnet/src/Agents/OpenAI/OpenAIAssistantChannel.cs
@@ -1,15 +1,8 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Runtime.CompilerServices;
-using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
-using Azure;
 using Azure.AI.OpenAI.Assistants;
-using Microsoft.Extensions.Logging;
-using Microsoft.SemanticKernel.ChatCompletion;
 
 namespace Microsoft.SemanticKernel.Agents.OpenAI;
 
@@ -19,25 +12,8 @@ namespace Microsoft.SemanticKernel.Agents.OpenAI;
 internal sealed class OpenAIAssistantChannel(AssistantsClient client, string threadId, OpenAIAssistantConfiguration.PollingConfiguration pollingConfiguration)
     : AgentChannel<OpenAIAssistantAgent>
 {
-    private const string FunctionDelimiter = "-";
-
-    private static readonly HashSet<RunStatus> s_pollingStatuses =
-        [
-            RunStatus.Queued,
-            RunStatus.InProgress,
-            RunStatus.Cancelling,
-        ];
-
-    private static readonly HashSet<RunStatus> s_terminalStatuses =
-        [
-            RunStatus.Expired,
-            RunStatus.Failed,
-            RunStatus.Cancelled,
-        ];
-
     private readonly AssistantsClient _client = client;
     private readonly string _threadId = threadId;
-    private readonly Dictionary<string, ToolDefinition[]> _agentTools = [];
 
     /// <inheritdoc/>
     protected override async Task ReceiveAsync(IReadOnlyList<ChatMessageContent> history, CancellationToken cancellationToken)
@@ -49,395 +25,18 @@ internal sealed class OpenAIAssistantChannel(AssistantsClient client, string thr
     }
 
     /// <inheritdoc/>
-    protected override async IAsyncEnumerable<ChatMessageContent> InvokeAsync(
+    protected override IAsyncEnumerable<ChatMessageContent> InvokeAsync(
         OpenAIAssistantAgent agent,
-        [EnumeratorCancellation] CancellationToken cancellationToken)
+        CancellationToken cancellationToken)
     {
-        if (agent.IsDeleted)
-        {
-            throw new KernelException($"Agent Failure - {nameof(OpenAIAssistantAgent)} agent is deleted: {agent.Id}.");
-        }
+        agent.ThrowIfDeleted();
 
-        if (!this._agentTools.TryGetValue(agent.Id, out ToolDefinition[]? tools))
-        {
-            tools = [.. agent.Tools, .. agent.Kernel.Plugins.SelectMany(p => p.Select(f => f.ToToolDefinition(p.Name, FunctionDelimiter)))];
-            this._agentTools.Add(agent.Id, tools);
-        }
-
-        this.Logger.LogDebug("[{MethodName}] Creating run for agent/thrad: {AgentId}/{ThreadId}", nameof(InvokeAsync), agent.Id, this._threadId);
-
-        CreateRunOptions options =
-            new(agent.Id)
-            {
-                OverrideInstructions = agent.Instructions,
-                OverrideTools = tools,
-            };
-
-        // Create run
-        ThreadRun run = await this._client.CreateRunAsync(this._threadId, options, cancellationToken).ConfigureAwait(false);
-
-        this.Logger.LogInformation("[{MethodName}] Created run: {RunId}", nameof(InvokeAsync), run.Id);
-
-        // Evaluate status and process steps and messages, as encountered.
-        HashSet<string> processedStepIds = [];
-        Dictionary<string, FunctionCallContent> functionSteps = [];
-
-        do
-        {
-            // Poll run and steps until actionable
-            PageableList<RunStep> steps = await PollRunStatusAsync().ConfigureAwait(false);
-
-            // Is in terminal state?
-            if (s_terminalStatuses.Contains(run.Status))
-            {
-                throw new KernelException($"Agent Failure - Run terminated: {run.Status} [{run.Id}]: {run.LastError?.Message ?? "Unknown"}");
-            }
-
-            // Is tool action required?
-            if (run.Status == RunStatus.RequiresAction)
-            {
-                this.Logger.LogDebug("[{MethodName}] Processing run steps: {RunId}", nameof(InvokeAsync), run.Id);
-
-                // Execute functions in parallel and post results at once.
-                FunctionCallContent[] activeFunctionSteps = steps.Data.SelectMany(step => ParseFunctionStep(agent, step)).ToArray();
-                if (activeFunctionSteps.Length > 0)
-                {
-                    // Emit function-call content
-                    yield return GenerateFunctionCallContent(agent.GetName(), activeFunctionSteps);
-
-                    // Invoke functions for each tool-step
-                    IEnumerable<Task<FunctionResultContent>> functionResultTasks = ExecuteFunctionSteps(agent, activeFunctionSteps, cancellationToken);
-
-                    // Block for function results
-                    FunctionResultContent[] functionResults = await Task.WhenAll(functionResultTasks).ConfigureAwait(false);
-
-                    // Process tool output
-                    ToolOutput[] toolOutputs = GenerateToolOutputs(functionResults);
-
-                    await this._client.SubmitToolOutputsToRunAsync(run, toolOutputs, cancellationToken).ConfigureAwait(false);
-                }
-
-                if (this.Logger.IsEnabled(LogLevel.Information)) // Avoid boxing if not enabled
-                {
-                    this.Logger.LogInformation("[{MethodName}] Processed #{MessageCount} run steps: {RunId}", nameof(InvokeAsync), activeFunctionSteps.Length, run.Id);
-                }
-            }
-
-            // Enumerate completed messages
-            this.Logger.LogDebug("[{MethodName}] Processing run messages: {RunId}", nameof(InvokeAsync), run.Id);
-
-            IEnumerable<RunStep> completedStepsToProcess =
-                steps
-                    .Where(s => s.CompletedAt.HasValue && !processedStepIds.Contains(s.Id))
-                    .OrderBy(s => s.CreatedAt);
-
-            int messageCount = 0;
-            foreach (RunStep completedStep in completedStepsToProcess)
-            {
-                if (completedStep.Type.Equals(RunStepType.ToolCalls))
-                {
-                    RunStepToolCallDetails toolCallDetails = (RunStepToolCallDetails)completedStep.StepDetails;
-
-                    foreach (RunStepToolCall toolCall in toolCallDetails.ToolCalls)
-                    {
-                        ChatMessageContent? content = null;
-
-                        // Process code-interpreter content
-                        if (toolCall is RunStepCodeInterpreterToolCall toolCodeInterpreter)
-                        {
-                            content = GenerateCodeInterpreterContent(agent.GetName(), toolCodeInterpreter);
-                        }
-                        // Process function result content
-                        else if (toolCall is RunStepFunctionToolCall toolFunction)
-                        {
-                            FunctionCallContent functionStep = functionSteps[toolFunction.Id]; // Function step always captured on invocation
-                            content = GenerateFunctionResultContent(agent.GetName(), functionStep, toolFunction.Output);
-                        }
-
-                        if (content is not null)
-                        {
-                            ++messageCount;
-
-                            yield return content;
-                        }
-                    }
-                }
-                else if (completedStep.Type.Equals(RunStepType.MessageCreation))
-                {
-                    RunStepMessageCreationDetails messageCreationDetails = (RunStepMessageCreationDetails)completedStep.StepDetails;
-
-                    // Retrieve the message
-                    ThreadMessage? message = await this.RetrieveMessageAsync(messageCreationDetails, cancellationToken).ConfigureAwait(false);
-
-                    if (message is not null)
-                    {
-                        AuthorRole role = new(message.Role.ToString());
-
-                        foreach (MessageContent itemContent in message.ContentItems)
-                        {
-                            ChatMessageContent? content = null;
-
-                            // Process text content
-                            if (itemContent is MessageTextContent contentMessage)
-                            {
-                                content = GenerateTextMessageContent(agent.GetName(), role, contentMessage);
-                            }
-                            // Process image content
-                            else if (itemContent is MessageImageFileContent contentImage)
-                            {
-                                content = GenerateImageFileContent(agent.GetName(), role, contentImage);
-                            }
-
-                            if (content is not null)
-                            {
-                                ++messageCount;
-
-                                yield return content;
-                            }
-                        }
-                    }
-                }
-
-                processedStepIds.Add(completedStep.Id);
-            }
-
-            if (this.Logger.IsEnabled(LogLevel.Information)) // Avoid boxing if not enabled
-            {
-                this.Logger.LogInformation("[{MethodName}] Processed #{MessageCount} run messages: {RunId}", nameof(InvokeAsync), messageCount, run.Id);
-            }
-        }
-        while (RunStatus.Completed != run.Status);
-
-        this.Logger.LogInformation("[{MethodName}] Completed run: {RunId}", nameof(InvokeAsync), run.Id);
-
-        // Local function to assist in run polling (participates in method closure).
-        async Task<PageableList<RunStep>> PollRunStatusAsync()
-        {
-            this.Logger.LogInformation("[{MethodName}] Polling run status: {RunId}", nameof(PollRunStatusAsync), run.Id);
-
-            int count = 0;
-
-            do
-            {
-                // Reduce polling frequency after a couple attempts
-                await Task.Delay(count >= 2 ? pollingConfiguration.RunPollingInterval : pollingConfiguration.RunPollingBackoff, cancellationToken).ConfigureAwait(false);
-                ++count;
-
-#pragma warning disable CA1031 // Do not catch general exception types
-                try
-                {
-                    run = await this._client.GetRunAsync(this._threadId, run.Id, cancellationToken).ConfigureAwait(false);
-                }
-                catch
-                {
-                    // Retry anyway..
-                }
-#pragma warning restore CA1031 // Do not catch general exception types
-            }
-            while (s_pollingStatuses.Contains(run.Status));
-
-            this.Logger.LogInformation("[{MethodName}] Run status is {RunStatus}: {RunId}", nameof(PollRunStatusAsync), run.Status, run.Id);
-
-            return await this._client.GetRunStepsAsync(run, cancellationToken: cancellationToken).ConfigureAwait(false);
-        }
-
-        // Local function to capture kernel function state for further processing (participates in method closure).
-        IEnumerable<FunctionCallContent> ParseFunctionStep(OpenAIAssistantAgent agent, RunStep step)
-        {
-            if (step.Status == RunStepStatus.InProgress && step.StepDetails is RunStepToolCallDetails callDetails)
-            {
-                foreach (RunStepFunctionToolCall toolCall in callDetails.ToolCalls.OfType<RunStepFunctionToolCall>())
-                {
-                    var nameParts = FunctionName.Parse(toolCall.Name, FunctionDelimiter);
-
-                    KernelArguments functionArguments = [];
-                    if (!string.IsNullOrWhiteSpace(toolCall.Arguments))
-                    {
-                        Dictionary<string, object> arguments = JsonSerializer.Deserialize<Dictionary<string, object>>(toolCall.Arguments)!;
-                        foreach (var argumentKvp in arguments)
-                        {
-                            functionArguments[argumentKvp.Key] = argumentKvp.Value.ToString();
-                        }
-                    }
-
-                    var content = new FunctionCallContent(nameParts.Name, nameParts.PluginName, toolCall.Id, functionArguments);
-
-                    functionSteps.Add(toolCall.Id, content);
-
-                    yield return content;
-                }
-            }
-        }
+        return OpenAIAssistantActions.InvokeAsync(agent, this._client, this._threadId, pollingConfiguration, this.Logger, cancellationToken); // %%% CONFIG
     }
 
     /// <inheritdoc/>
     protected override IAsyncEnumerable<ChatMessageContent> GetHistoryAsync(CancellationToken cancellationToken)
     {
         return OpenAIAssistantActions.GetMessagesAsync(this._client, this._threadId, cancellationToken);
-    }
-
-    private static AnnotationContent GenerateAnnotationContent(MessageTextAnnotation annotation)
-    {
-        string? fileId = null;
-        if (annotation is MessageTextFileCitationAnnotation citationAnnotation)
-        {
-            fileId = citationAnnotation.FileId;
-        }
-        else if (annotation is MessageTextFilePathAnnotation pathAnnotation)
-        {
-            fileId = pathAnnotation.FileId;
-        }
-
-        return
-            new()
-            {
-                Quote = annotation.Text,
-                StartIndex = annotation.StartIndex,
-                EndIndex = annotation.EndIndex,
-                FileId = fileId,
-            };
-    }
-
-    private static ChatMessageContent GenerateImageFileContent(string agentName, AuthorRole role, MessageImageFileContent contentImage)
-    {
-        return
-            new ChatMessageContent(
-                role,
-                [
-                    new FileReferenceContent(contentImage.FileId)
-                ])
-            {
-                AuthorName = agentName,
-            };
-    }
-
-    private static ChatMessageContent? GenerateTextMessageContent(string agentName, AuthorRole role, MessageTextContent contentMessage)
-    {
-        ChatMessageContent? messageContent = null;
-
-        string textContent = contentMessage.Text.Trim();
-
-        if (!string.IsNullOrWhiteSpace(textContent))
-        {
-            messageContent =
-                new(role, textContent)
-                {
-                    AuthorName = agentName
-                };
-
-            foreach (MessageTextAnnotation annotation in contentMessage.Annotations)
-            {
-                messageContent.Items.Add(GenerateAnnotationContent(annotation));
-            }
-        }
-
-        return messageContent;
-    }
-
-    private static ChatMessageContent GenerateCodeInterpreterContent(string agentName, RunStepCodeInterpreterToolCall contentCodeInterpreter)
-    {
-        return
-            new ChatMessageContent(
-                AuthorRole.Tool,
-                [
-                    new TextContent(contentCodeInterpreter.Input)
-                ])
-            {
-                AuthorName = agentName,
-            };
-    }
-
-    private static ChatMessageContent GenerateFunctionCallContent(string agentName, FunctionCallContent[] functionSteps)
-    {
-        ChatMessageContent functionCallContent = new(AuthorRole.Tool, content: null)
-        {
-            AuthorName = agentName
-        };
-
-        functionCallContent.Items.AddRange(functionSteps);
-
-        return functionCallContent;
-    }
-
-    private static ChatMessageContent GenerateFunctionResultContent(string agentName, FunctionCallContent functionStep, string result)
-    {
-        ChatMessageContent functionCallContent = new(AuthorRole.Tool, content: null)
-        {
-            AuthorName = agentName
-        };
-
-        functionCallContent.Items.Add(
-            new FunctionResultContent(
-                functionStep.FunctionName,
-                functionStep.PluginName,
-                functionStep.Id,
-                result));
-
-        return functionCallContent;
-    }
-
-    private static Task<FunctionResultContent>[] ExecuteFunctionSteps(OpenAIAssistantAgent agent, FunctionCallContent[] functionSteps, CancellationToken cancellationToken)
-    {
-        Task<FunctionResultContent>[] functionTasks = new Task<FunctionResultContent>[functionSteps.Length];
-
-        for (int index = 0; index < functionSteps.Length; ++index)
-        {
-            functionTasks[index] = functionSteps[index].InvokeAsync(agent.Kernel, cancellationToken);
-        }
-
-        return functionTasks;
-    }
-
-    private static ToolOutput[] GenerateToolOutputs(FunctionResultContent[] functionResults)
-    {
-        ToolOutput[] toolOutputs = new ToolOutput[functionResults.Length];
-
-        for (int index = 0; index < functionResults.Length; ++index)
-        {
-            FunctionResultContent functionResult = functionResults[index];
-
-            object resultValue = (functionResult.Result as FunctionResult)?.GetValue<object>() ?? string.Empty;
-
-            if (resultValue is not string textResult)
-            {
-                textResult = JsonSerializer.Serialize(resultValue);
-            }
-
-            toolOutputs[index] = new ToolOutput(functionResult.CallId, textResult!);
-        }
-
-        return toolOutputs;
-    }
-
-    private async Task<ThreadMessage?> RetrieveMessageAsync(RunStepMessageCreationDetails detail, CancellationToken cancellationToken)
-    {
-        ThreadMessage? message = null;
-
-        bool retry = false;
-        int count = 0;
-        do
-        {
-            try
-            {
-                message = await this._client.GetMessageAsync(this._threadId, detail.MessageCreation.MessageId, cancellationToken).ConfigureAwait(false);
-            }
-            catch (RequestFailedException exception)
-            {
-                // Step has provided the message-id.  Retry on of NotFound/404 exists.
-                // Extremely rarely there might be a synchronization issue between the
-                // assistant response and message-service.
-                retry = exception.Status == (int)HttpStatusCode.NotFound && count < 3;
-            }
-
-            if (retry)
-            {
-                await Task.Delay(pollingConfiguration.MessageSynchronizationDelay, cancellationToken).ConfigureAwait(false);
-            }
-
-            ++count;
-        }
-        while (retry);
-
-        return message;
     }
 }
